deleted docks cuz it was casuing issues and added new colmuns with 0s as default
