--- CargoHubV2/alembic/versions/18a2dd8935ed_orders8.py
@@ -1,339 +1,0 @@
-"""Orders8
-
-Revision ID: 18a2dd8935ed
-Revises: 
-Create Date: 2024-11-22 23:46:50.297130
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import sqlite
-
-# revision identifiers, used by Alembic.
-revision: str = '18a2dd8935ed'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('api_keys',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('key', sa.String(), nullable=False),
-    sa.Column('role', sa.String(), nullable=False),
-    sa.Column('permissions', sqlite.JSON(), nullable=False),
-    sa.Column('is_active', sa.Boolean(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('key')
-    )
-    op.create_table('clients',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=True),
-    sa.Column('address', sa.String(), nullable=True),
-    sa.Column('city', sa.String(), nullable=True),
-    sa.Column('zip_code', sa.String(), nullable=True),
-    sa.Column('province', sa.String(), nullable=True),
-    sa.Column('country', sa.String(), nullable=True),
-    sa.Column('contact_name', sa.String(), nullable=True),
-    sa.Column('contact_phone', sa.String(), nullable=True),
-    sa.Column('contact_email', sa.String(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_clients_address'), 'clients', ['address'], unique=False)
-    op.create_index(op.f('ix_clients_city'), 'clients', ['city'], unique=False)
-    op.create_index(op.f('ix_clients_contact_email'), 'clients', ['contact_email'], unique=False)
-    op.create_index(op.f('ix_clients_contact_name'), 'clients', ['contact_name'], unique=False)
-    op.create_index(op.f('ix_clients_contact_phone'), 'clients', ['contact_phone'], unique=False)
-    op.create_index(op.f('ix_clients_country'), 'clients', ['country'], unique=False)
-    op.create_index(op.f('ix_clients_created_at'), 'clients', ['created_at'], unique=False)
-    op.create_index(op.f('ix_clients_id'), 'clients', ['id'], unique=False)
-    op.create_index(op.f('ix_clients_name'), 'clients', ['name'], unique=False)
-    op.create_index(op.f('ix_clients_province'), 'clients', ['province'], unique=False)
-    op.create_index(op.f('ix_clients_updated_at'), 'clients', ['updated_at'], unique=False)
-    op.create_index(op.f('ix_clients_zip_code'), 'clients', ['zip_code'], unique=False)
-    op.create_table('item_groups',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('name', sa.String(), nullable=True),
-    sa.Column('description', sa.String(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_item_groups_name'), 'item_groups', ['name'], unique=True)
-    op.create_table('item_lines',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('name', sa.String(), nullable=True),
-    sa.Column('description', sa.String(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_item_lines_name'), 'item_lines', ['name'], unique=True)
-    op.create_table('item_types',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('name', sa.String(), nullable=True),
-    sa.Column('description', sa.String(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_item_types_name'), 'item_types', ['name'], unique=True)
-    op.create_table('locations',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('warehouse_id', sa.Integer(), nullable=True),
-    sa.Column('code', sa.String(), nullable=True),
-    sa.Column('name', sa.String(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_locations_code'), 'locations', ['code'], unique=False)
-    op.create_index(op.f('ix_locations_created_at'), 'locations', ['created_at'], unique=False)
-    op.create_index(op.f('ix_locations_id'), 'locations', ['id'], unique=False)
-    op.create_index(op.f('ix_locations_name'), 'locations', ['name'], unique=False)
-    op.create_index(op.f('ix_locations_updated_at'), 'locations', ['updated_at'], unique=False)
-    op.create_index(op.f('ix_locations_warehouse_id'), 'locations', ['warehouse_id'], unique=False)
-    op.create_table('shipments',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('order_id', sa.Integer(), nullable=True),
-    sa.Column('source_id', sa.Integer(), nullable=True),
-    sa.Column('order_date', sa.DateTime(), nullable=True),
-    sa.Column('request_date', sa.DateTime(), nullable=True),
-    sa.Column('shipment_date', sa.DateTime(), nullable=True),
-    sa.Column('shipment_type', sa.String(), nullable=True),
-    sa.Column('shipment_status', sa.String(), nullable=True),
-    sa.Column('notes', sa.String(), nullable=True),
-    sa.Column('carrier_code', sa.String(), nullable=True),
-    sa.Column('carrier_description', sa.String(), nullable=True),
-    sa.Column('service_code', sa.String(), nullable=True),
-    sa.Column('payment_type', sa.String(), nullable=True),
-    sa.Column('transfer_mode', sa.String(), nullable=True),
-    sa.Column('total_package_count', sa.Integer(), nullable=True),
-    sa.Column('total_package_weight', sa.Float(), nullable=True),
-    sa.Column('items', sa.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_shipments_carrier_code'), 'shipments', ['carrier_code'], unique=False)
-    op.create_index(op.f('ix_shipments_carrier_description'), 'shipments', ['carrier_description'], unique=False)
-    op.create_index(op.f('ix_shipments_id'), 'shipments', ['id'], unique=False)
-    op.create_index(op.f('ix_shipments_order_date'), 'shipments', ['order_date'], unique=False)
-    op.create_index(op.f('ix_shipments_order_id'), 'shipments', ['order_id'], unique=False)
-    op.create_index(op.f('ix_shipments_payment_type'), 'shipments', ['payment_type'], unique=False)
-    op.create_index(op.f('ix_shipments_request_date'), 'shipments', ['request_date'], unique=False)
-    op.create_index(op.f('ix_shipments_service_code'), 'shipments', ['service_code'], unique=False)
-    op.create_index(op.f('ix_shipments_shipment_date'), 'shipments', ['shipment_date'], unique=False)
-    op.create_index(op.f('ix_shipments_shipment_status'), 'shipments', ['shipment_status'], unique=False)
-    op.create_index(op.f('ix_shipments_shipment_type'), 'shipments', ['shipment_type'], unique=False)
-    op.create_index(op.f('ix_shipments_source_id'), 'shipments', ['source_id'], unique=False)
-    op.create_index(op.f('ix_shipments_transfer_mode'), 'shipments', ['transfer_mode'], unique=False)
-    op.create_table('suppliers',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('code', sa.String(), nullable=True),
-    sa.Column('name', sa.String(), nullable=True),
-    sa.Column('address', sa.String(), nullable=True),
-    sa.Column('address_extra', sa.String(), nullable=True),
-    sa.Column('city', sa.String(), nullable=True),
-    sa.Column('zip_code', sa.String(), nullable=True),
-    sa.Column('province', sa.String(), nullable=True),
-    sa.Column('country', sa.String(), nullable=True),
-    sa.Column('contact_name', sa.String(), nullable=True),
-    sa.Column('phonenumber', sa.String(), nullable=True),
-    sa.Column('reference', sa.String(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_suppliers_code'), 'suppliers', ['code'], unique=True)
-    op.create_index(op.f('ix_suppliers_id'), 'suppliers', ['id'], unique=False)
-    op.create_table('transfers',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('reference', sa.String(), nullable=True),
-    sa.Column('transfer_from', sa.Integer(), nullable=True),
-    sa.Column('transfer_to', sa.Integer(), nullable=False),
-    sa.Column('transfer_status', sa.String(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.Column('items', sa.JSON(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_transfers_created_at'), 'transfers', ['created_at'], unique=False)
-    op.create_index(op.f('ix_transfers_id'), 'transfers', ['id'], unique=False)
-    op.create_index(op.f('ix_transfers_items'), 'transfers', ['items'], unique=False)
-    op.create_index(op.f('ix_transfers_reference'), 'transfers', ['reference'], unique=True)
-    op.create_index(op.f('ix_transfers_transfer_status'), 'transfers', ['transfer_status'], unique=False)
-    op.create_index(op.f('ix_transfers_updated_at'), 'transfers', ['updated_at'], unique=False)
-    op.create_table('warehouses',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('code', sa.String(), nullable=True),
-    sa.Column('name', sa.String(), nullable=True),
-    sa.Column('address', sa.String(), nullable=True),
-    sa.Column('zip', sa.String(), nullable=True),
-    sa.Column('city', sa.String(), nullable=True),
-    sa.Column('province', sa.String(), nullable=True),
-    sa.Column('country', sa.String(), nullable=True),
-    sa.Column('contact', sa.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_warehouses_address'), 'warehouses', ['address'], unique=False)
-    op.create_index(op.f('ix_warehouses_city'), 'warehouses', ['city'], unique=False)
-    op.create_index(op.f('ix_warehouses_code'), 'warehouses', ['code'], unique=True)
-    op.create_index(op.f('ix_warehouses_contact'), 'warehouses', ['contact'], unique=False)
-    op.create_index(op.f('ix_warehouses_country'), 'warehouses', ['country'], unique=False)
-    op.create_index(op.f('ix_warehouses_id'), 'warehouses', ['id'], unique=False)
-    op.create_index(op.f('ix_warehouses_name'), 'warehouses', ['name'], unique=False)
-    op.create_index(op.f('ix_warehouses_province'), 'warehouses', ['province'], unique=False)
-    op.create_index(op.f('ix_warehouses_zip'), 'warehouses', ['zip'], unique=False)
-    op.create_table('items',
-    sa.Column('uid', sa.String(), nullable=False),
-    sa.Column('code', sa.String(), nullable=True),
-    sa.Column('description', sa.String(), nullable=True),
-    sa.Column('short_description', sa.String(), nullable=True),
-    sa.Column('upc_code', sa.String(), nullable=True),
-    sa.Column('model_number', sa.String(), nullable=True),
-    sa.Column('commodity_code', sa.String(), nullable=True),
-    sa.Column('item_line', sa.Integer(), nullable=True),
-    sa.Column('item_group', sa.Integer(), nullable=True),
-    sa.Column('item_type', sa.Integer(), nullable=True),
-    sa.Column('unit_purchase_quantity', sa.Integer(), nullable=True),
-    sa.Column('unit_order_quantity', sa.Integer(), nullable=True),
-    sa.Column('pack_order_quantity', sa.Integer(), nullable=True),
-    sa.Column('supplier_id', sa.Integer(), nullable=True),
-    sa.Column('supplier_code', sa.String(), nullable=True),
-    sa.Column('supplier_part_number', sa.String(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['item_group'], ['item_groups.id'], ),
-    sa.ForeignKeyConstraint(['item_line'], ['item_lines.id'], ),
-    sa.ForeignKeyConstraint(['item_type'], ['item_types.id'], ),
-    sa.ForeignKeyConstraint(['supplier_id'], ['suppliers.id'], ),
-    sa.PrimaryKeyConstraint('uid')
-    )
-    op.create_index(op.f('ix_items_code'), 'items', ['code'], unique=True)
-    op.create_index(op.f('ix_items_uid'), 'items', ['uid'], unique=False)
-    op.create_table('orders',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('source_id', sa.Integer(), nullable=False),
-    sa.Column('order_date', sa.DateTime(), nullable=False),
-    sa.Column('request_date', sa.DateTime(), nullable=True),
-    sa.Column('reference', sa.String(), nullable=False),
-    sa.Column('reference_extra', sa.String(), nullable=True),
-    sa.Column('order_status', sa.String(), nullable=False),
-    sa.Column('notes', sa.String(), nullable=True),
-    sa.Column('shipping_notes', sa.String(), nullable=True),
-    sa.Column('picking_notes', sa.String(), nullable=True),
-    sa.Column('warehouse_id', sa.Integer(), nullable=False),
-    sa.Column('ship_to', sa.Integer(), nullable=True),
-    sa.Column('bill_to', sa.Integer(), nullable=True),
-    sa.Column('shipment_id', sa.Integer(), nullable=True),
-    sa.Column('total_amount', sa.Float(), nullable=False),
-    sa.Column('total_discount', sa.Float(), nullable=True),
-    sa.Column('total_tax', sa.Float(), nullable=True),
-    sa.Column('total_surcharge', sa.Float(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.Column('items', sa.JSON(), nullable=True),
-    sa.ForeignKeyConstraint(['warehouse_id'], ['warehouses.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_orders_id'), 'orders', ['id'], unique=False)
-    op.create_table('inventories',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('item_id', sa.Integer(), nullable=False),
-    sa.Column('description', sa.String(), nullable=True),
-    sa.Column('item_reference', sa.String(), nullable=True),
-    sa.Column('locations', sa.JSON(), nullable=True),
-    sa.Column('total_on_hand', sa.Integer(), nullable=True),
-    sa.Column('total_expected', sa.Integer(), nullable=True),
-    sa.Column('total_ordered', sa.Integer(), nullable=True),
-    sa.Column('total_allocated', sa.Integer(), nullable=True),
-    sa.Column('total_available', sa.Integer(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['item_id'], ['items.uid'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('item_id'),
-    sa.UniqueConstraint('item_reference')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('inventories')
-    op.drop_index(op.f('ix_orders_id'), table_name='orders')
-    op.drop_table('orders')
-    op.drop_index(op.f('ix_items_uid'), table_name='items')
-    op.drop_index(op.f('ix_items_code'), table_name='items')
-    op.drop_table('items')
-    op.drop_index(op.f('ix_warehouses_zip'), table_name='warehouses')
-    op.drop_index(op.f('ix_warehouses_province'), table_name='warehouses')
-    op.drop_index(op.f('ix_warehouses_name'), table_name='warehouses')
-    op.drop_index(op.f('ix_warehouses_id'), table_name='warehouses')
-    op.drop_index(op.f('ix_warehouses_country'), table_name='warehouses')
-    op.drop_index(op.f('ix_warehouses_contact'), table_name='warehouses')
-    op.drop_index(op.f('ix_warehouses_code'), table_name='warehouses')
-    op.drop_index(op.f('ix_warehouses_city'), table_name='warehouses')
-    op.drop_index(op.f('ix_warehouses_address'), table_name='warehouses')
-    op.drop_table('warehouses')
-    op.drop_index(op.f('ix_transfers_updated_at'), table_name='transfers')
-    op.drop_index(op.f('ix_transfers_transfer_status'), table_name='transfers')
-    op.drop_index(op.f('ix_transfers_reference'), table_name='transfers')
-    op.drop_index(op.f('ix_transfers_items'), table_name='transfers')
-    op.drop_index(op.f('ix_transfers_id'), table_name='transfers')
-    op.drop_index(op.f('ix_transfers_created_at'), table_name='transfers')
-    op.drop_table('transfers')
-    op.drop_index(op.f('ix_suppliers_id'), table_name='suppliers')
-    op.drop_index(op.f('ix_suppliers_code'), table_name='suppliers')
-    op.drop_table('suppliers')
-    op.drop_index(op.f('ix_shipments_transfer_mode'), table_name='shipments')
-    op.drop_index(op.f('ix_shipments_source_id'), table_name='shipments')
-    op.drop_index(op.f('ix_shipments_shipment_type'), table_name='shipments')
-    op.drop_index(op.f('ix_shipments_shipment_status'), table_name='shipments')
-    op.drop_index(op.f('ix_shipments_shipment_date'), table_name='shipments')
-    op.drop_index(op.f('ix_shipments_service_code'), table_name='shipments')
-    op.drop_index(op.f('ix_shipments_request_date'), table_name='shipments')
-    op.drop_index(op.f('ix_shipments_payment_type'), table_name='shipments')
-    op.drop_index(op.f('ix_shipments_order_id'), table_name='shipments')
-    op.drop_index(op.f('ix_shipments_order_date'), table_name='shipments')
-    op.drop_index(op.f('ix_shipments_id'), table_name='shipments')
-    op.drop_index(op.f('ix_shipments_carrier_description'), table_name='shipments')
-    op.drop_index(op.f('ix_shipments_carrier_code'), table_name='shipments')
-    op.drop_table('shipments')
-    op.drop_index(op.f('ix_locations_warehouse_id'), table_name='locations')
-    op.drop_index(op.f('ix_locations_updated_at'), table_name='locations')
-    op.drop_index(op.f('ix_locations_name'), table_name='locations')
-    op.drop_index(op.f('ix_locations_id'), table_name='locations')
-    op.drop_index(op.f('ix_locations_created_at'), table_name='locations')
-    op.drop_index(op.f('ix_locations_code'), table_name='locations')
-    op.drop_table('locations')
-    op.drop_index(op.f('ix_item_types_name'), table_name='item_types')
-    op.drop_table('item_types')
-    op.drop_index(op.f('ix_item_lines_name'), table_name='item_lines')
-    op.drop_table('item_lines')
-    op.drop_index(op.f('ix_item_groups_name'), table_name='item_groups')
-    op.drop_table('item_groups')
-    op.drop_index(op.f('ix_clients_zip_code'), table_name='clients')
-    op.drop_index(op.f('ix_clients_updated_at'), table_name='clients')
-    op.drop_index(op.f('ix_clients_province'), table_name='clients')
-    op.drop_index(op.f('ix_clients_name'), table_name='clients')
-    op.drop_index(op.f('ix_clients_id'), table_name='clients')
-    op.drop_index(op.f('ix_clients_created_at'), table_name='clients')
-    op.drop_index(op.f('ix_clients_country'), table_name='clients')
-    op.drop_index(op.f('ix_clients_contact_phone'), table_name='clients')
-    op.drop_index(op.f('ix_clients_contact_name'), table_name='clients')
-    op.drop_index(op.f('ix_clients_contact_email'), table_name='clients')
-    op.drop_index(op.f('ix_clients_city'), table_name='clients')
-    op.drop_index(op.f('ix_clients_address'), table_name='clients')
-    op.drop_table('clients')
-    op.drop_table('api_keys')
-    # ### end Alembic commands ###
